# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Enterprise Ticketing, Project Management & Workflow System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tickets:
    get:
      operationId: listTickets
      tags: [tickets]
      summary: List tickets
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: priority
          in: query
          schema: { type: string }
        - name: assignedTo
          in: query
          schema: { type: integer }
        - name: projectId
          in: query
          schema: { type: integer }
        - name: category
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: fromDate
          in: query
          schema: { type: string }
        - name: toDate
          in: query
          schema: { type: string }
        - name: myTickets
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
    post:
      operationId: createTicket
      tags: [tickets]
      summary: Create ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}:
    get:
      operationId: getTicket
      tags: [tickets]
      summary: Get ticket
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Ticket
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
    patch:
      operationId: updateTicket
      tags: [tickets]
      summary: Update ticket
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
    delete:
      operationId: deleteTicket
      tags: [tickets]
      summary: Delete ticket
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /tickets/{id}/status:
    patch:
      operationId: updateTicketStatus
      tags: [tickets]
      summary: Update ticket status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}/assign:
    patch:
      operationId: assignTicket
      tags: [tickets]
      summary: Assign ticket
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketAssign"
      responses:
        "200":
          description: Assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}/forward:
    post:
      operationId: forwardTicket
      tags: [tickets]
      summary: Forward ticket
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TicketForward"
      responses:
        "200":
          description: Forwarded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
  /tickets/{id}/comments:
    get:
      operationId: listTicketComments
      tags: [tickets]
      summary: List ticket comments
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: addTicketComment
      tags: [tickets]
      summary: Add comment to ticket
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"

  # ── WORKLIST ─────────────────────────────────────────────────────────────────
  /worklist:
    get:
      operationId: getWorklist
      tags: [worklist]
      summary: Get common worklist (unassigned tickets pool)
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: priority
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Worklist tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
  /worklist/{id}/pick:
    post:
      operationId: pickTicket
      tags: [worklist]
      summary: Pick a ticket from the worklist (self-assign)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Picked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"

  # ── PROJECTS ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Ticket:
      type: object
      required: [id, ticketNo, subject, status, priority, createdAt]
      properties:
        id: { type: integer }
        ticketNo: { type: string }
        subject: { type: string }
        description: { type: ["string", "null"] }
        status: { type: string }
        priority: { type: string }
        category: { type: ["string", "null"] }
        categoryId: { type: ["integer", "null"] }
        type: { type: string }
        createdById: { type: integer }
        createdByName: { type: ["string", "null"] }
        assignedToId: { type: ["integer", "null"] }
        assignedToName: { type: ["string", "null"] }
        assignedToAvatar: { type: ["string", "null"] }
        projectId: { type: ["integer", "null"] }
        projectName: { type: ["string", "null"] }
        dueDate: { type: ["string", "null"] }
        resolvedAt: { type: ["string", "null"] }
        closedAt: { type: ["string", "null"] }
        pendingDays: { type: ["integer", "null"] }
        slaBreached: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
    TicketInput:
      type: object
      required: [subject, priority, type]
      properties:
        subject: { type: string }
        description: { type: string }
        priority: { type: string }
        categoryId: { type: integer }
        type: { type: string }
        assignedToId: { type: integer }
        projectId: { type: integer }
        dueDate: { type: string }
    TicketUpdate:
      type: object
      properties:
        subject: { type: string }
        description: { type: string }
        priority: { type: string }
        categoryId: { type: integer }
        dueDate: { type: string }
    TicketStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { type: string }
        remarks: { type: string }
    TicketAssign:
      type: object
      required: [assignedToId]
      properties:
        assignedToId: { type: integer }
        remarks: { type: string }
    TicketForward:
      type: object
      required: [forwardToId]
      properties:
        forwardToId: { type: integer }
        remarks: { type: string }
    Comment:
      type: object
      required: [id, content, authorId, createdAt]
      properties:
        id: { type: integer }
        content: { type: string }
        authorId: { type: integer }
        authorName: { type: ["string", "null"] }
        authorAvatar: { type: ["string", "null"] }
        createdAt: { type: string }
    CommentInput:
      type: object
      required: [content]
      properties:
        content: { type: string }

    # ── PROJECT SCHEMAS ─────────────────────────────────────────────────────
...